+++ snowflake/ml/monitoring/_client/model_monitor_sql_client.py
@@ -1,6 +1,4 @@
-import typing
-from collections import Counter
-from typing import Any, Dict, List, Mapping, Optional, Set
+from typing import Any, Dict, List, Mapping, Optional
 
 from snowflake import snowpark
 from snowflake.ml._internal.utils import (
@@ -10,26 +8,11 @@ from snowflake.ml._internal.utils import (
     table_manager,
 )
 from snowflake.ml.model._client.sql import _base
-from snowflake.ml.model._model_composer.model_manifest import model_manifest_schema
 from snowflake.snowpark import session, types
-
-SNOWML_MONITORING_METADATA_TABLE_NAME = "_SYSTEM_MONITORING_METADATA"
 
 MODEL_JSON_COL_NAME = "model"
 MODEL_JSON_MODEL_NAME_FIELD = "model_name"
 MODEL_JSON_VERSION_NAME_FIELD = "version_name"
-
-MONITOR_NAME_COL_NAME = "MONITOR_NAME"
-SOURCE_TABLE_NAME_COL_NAME = "SOURCE_TABLE_NAME"
-FQ_MODEL_NAME_COL_NAME = "FULLY_QUALIFIED_MODEL_NAME"
-VERSION_NAME_COL_NAME = "MODEL_VERSION_NAME"
-FUNCTION_NAME_COL_NAME = "FUNCTION_NAME"
-TASK_COL_NAME = "TASK"
-MONITORING_ENABLED_COL_NAME = "IS_ENABLED"
-TIMESTAMP_COL_NAME_COL_NAME = "TIMESTAMP_COLUMN_NAME"
-PREDICTION_COL_NAMES_COL_NAME = "PREDICTION_COLUMN_NAMES"
-LABEL_COL_NAMES_COL_NAME = "LABEL_COLUMN_NAMES"
-ID_COL_NAMES_COL_NAME = "ID_COLUMN_NAMES"
 
 
 def _build_sql_list_from_columns(columns: List[sql_identifier.SqlIdentifier]) -> str:
@@ -146,19 +129,6 @@ class ModelMonitorSQLClient:
             .validate()
         )
 
-    def _validate_unique_columns(
-        self,
-        timestamp_column: sql_identifier.SqlIdentifier,
-        id_columns: List[sql_identifier.SqlIdentifier],
-        prediction_columns: List[sql_identifier.SqlIdentifier],
-        label_columns: List[sql_identifier.SqlIdentifier],
-    ) -> None:
-        all_columns = [*id_columns, *prediction_columns, *label_columns, timestamp_column]
-        num_all_columns = len(all_columns)
-        num_unique_columns = len(set(all_columns))
-        if num_all_columns != num_unique_columns:
-            raise ValueError("Column names must be unique across id, timestamp, prediction, and label columns.")
-
     def validate_existence_by_name(
         self,
         *,
@@ -244,125 +214,6 @@ class ModelMonitorSQLClient:
         if not all([column_name in source_column_schema for column_name in id_columns]):
             raise ValueError(f"ID column(s): {id_columns} do not exist in source.")
 
-    def _validate_timestamp_column_type(
-        self, table_schema: Mapping[str, types.DataType], timestamp_column: sql_identifier.SqlIdentifier
-    ) -> None:
-        """Ensures columns have the same type.
-
-        Args:
-            table_schema: Dictionary of column names and types in the source table.
-            timestamp_column: Name of the timestamp column.
-
-        Raises:
-            ValueError: If the timestamp column is not of type TimestampType.
-        """
-        if not isinstance(table_schema[timestamp_column], types.TimestampType):
-            raise ValueError(
-                f"Timestamp column: {timestamp_column} must be TimestampType. "
-                f"Found: {table_schema[timestamp_column]}"
-            )
-
-    def _validate_id_columns_types(
-        self, table_schema: Mapping[str, types.DataType], id_columns: List[sql_identifier.SqlIdentifier]
-    ) -> None:
-        """Ensures id columns have the correct type.
-
-        Args:
-            table_schema: Dictionary of column names and types in the source table.
-            id_columns: List of id column names.
-
-        Raises:
-            ValueError: If the id column is not of type StringType.
-        """
-        id_column_types = list({table_schema[column_name] for column_name in id_columns})
-        all_id_columns_string = all([isinstance(column_type, types.StringType) for column_type in id_column_types])
-        if not all_id_columns_string:
-            raise ValueError(f"Id columns must all be StringType. Found: {id_column_types}")
-
-    def _validate_prediction_columns_types(
-        self, table_schema: Mapping[str, types.DataType], prediction_columns: List[sql_identifier.SqlIdentifier]
-    ) -> None:
-        """Ensures prediction columns have the same type.
-
-        Args:
-            table_schema: Dictionary of column names and types in the source table.
-            prediction_columns: List of prediction column names.
-
-        Raises:
-            ValueError: If the prediction columns do not share the same type.
-        """
-
-        prediction_column_types = {table_schema[column_name] for column_name in prediction_columns}
-        if len(prediction_column_types) > 1:
-            raise ValueError(f"Prediction column types must be the same. Found: {prediction_column_types}")
-
-    def _validate_label_columns_types(
-        self,
-        table_schema: Mapping[str, types.DataType],
-        label_columns: List[sql_identifier.SqlIdentifier],
-    ) -> None:
-        """Ensures label columns have the same type, and the correct type for the score type.
-
-        Args:
-            table_schema: Dictionary of column names and types in the source table.
-            label_columns: List of label column names.
-
-        Raises:
-            ValueError: If the label columns do not share the same type.
-        """
-        label_column_types = {table_schema[column_name] for column_name in label_columns}
-        if len(label_column_types) > 1:
-            raise ValueError(f"Label column types must be the same. Found: {label_column_types}")
-
-    def _validate_column_types(
-        self,
-        *,
-        table_schema: Mapping[str, types.DataType],
-        timestamp_column: sql_identifier.SqlIdentifier,
-        id_columns: List[sql_identifier.SqlIdentifier],
-        prediction_columns: List[sql_identifier.SqlIdentifier],
-        label_columns: List[sql_identifier.SqlIdentifier],
-    ) -> None:
-        """Ensures columns have the expected type.
-
-        Args:
-            table_schema: Dictionary of column names and types in the source table.
-            timestamp_column: Name of the timestamp column.
-            id_columns: List of id column names.
-            prediction_columns: List of prediction column names.
-            label_columns: List of label column names.
-        """
-        self._validate_timestamp_column_type(table_schema, timestamp_column)
-        self._validate_id_columns_types(table_schema, id_columns)
-        self._validate_prediction_columns_types(table_schema, prediction_columns)
-        self._validate_label_columns_types(table_schema, label_columns)
-        # TODO(SNOW-1646693): Validate label makes sense with model task
-
-    def _validate_source_table_features_shape(
-        self,
-        table_schema: Mapping[str, types.DataType],
-        special_columns: Set[sql_identifier.SqlIdentifier],
-        model_function: model_manifest_schema.ModelFunctionInfo,
-    ) -> None:
-        table_schema_without_special_columns = {
-            k: v for k, v in table_schema.items() if sql_identifier.SqlIdentifier(k) not in special_columns
-        }
-        schema_column_types_to_count: typing.Counter[types.DataType] = Counter()
-        for column_type in table_schema_without_special_columns.values():
-            schema_column_types_to_count[column_type] += 1
-
-        inputs = model_function["signature"].inputs
-        function_input_types = [input.as_snowpark_type() for input in inputs]
-        function_input_types_to_count: typing.Counter[types.DataType] = Counter()
-        for function_input_type in function_input_types:
-            function_input_types_to_count[function_input_type] += 1
-
-        if function_input_types_to_count != schema_column_types_to_count:
-            raise ValueError(
-                "Model function input types do not match the source table input columns types. "
-                f"Model function expected: {inputs} but got {table_schema_without_special_columns}"
-            )
-
     def validate_source(
         self,
         *,
@@ -394,22 +245,6 @@ class ModelMonitorSQLClient:
             actual_class_columns=actual_class_columns,
             id_columns=id_columns,
         )
-
-    def delete_monitor_metadata(
-        self,
-        name: str,
-        statement_params: Optional[Dict[str, Any]] = None,
-    ) -> None:
-        """Delete the row in the metadata table corresponding to the given monitor name.
-
-        Args:
-            name: Name of the model monitor whose metadata should be deleted.
-            statement_params: Optional set of statement_params to include with query.
-        """
-        self._sql_client._session.sql(
-            f"""DELETE FROM {self._database_name}.{self._schema_name}.{SNOWML_MONITORING_METADATA_TABLE_NAME}
-            WHERE {MONITOR_NAME_COL_NAME} = '{name}'""",
-        ).collect(statement_params=statement_params)
 
     def _alter_monitor(
         self,